KAA-292: done refactoring

+++ client/client-multi/client-objective-c/Kaa/KaaClientProperties.h
@@ -43,9 +43,9 @@
 
 - (NSString *)sdkToken;
 
-- (NSInteger)pollDelay;
+- (int32_t)pollDelay;
 
-- (NSInteger)pollPeriod;
+- (int32_t)pollPeriod;
 
 - (TimeUnit)pollUnit;
