+++ client/client-multi/client-objective-c/Kaa/context/AbstractExecutorContext.h
@@ -20,10 +20,10 @@
 
 @interface AbstractExecutorContext : NSObject <ExecutorContext>
 
-@property (nonatomic) NSInteger timeOut;
+@property (nonatomic) int64_t timeOut;
 @property (nonatomic) TimeUnit timeUnit;
 
-- (instancetype)initWithTimeOut:(NSInteger)timeOut andTimeUnit:(TimeUnit)timeUnit;
+- (instancetype)initWithTimeOut:(int64_t)timeOut andTimeUnit:(TimeUnit)timeUnit;
 - (void)shutDownExecutor:(NSOperationQueue*)queue;
 
 @end
